backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tinkerpilot"
version = "0.1.0"
description = "Local AI assistant for developers - privacy-first, offline productivity"
requires-python = ">=3.10"
dependencies = [
    # AI / ML
    "llama-cpp-python>=0.3.0",
    "faster-whisper>=1.0.0",
    "chromadb>=0.5.0",
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "python-multipart>=0.0.9",
    # CLI
    "typer[all]>=0.12.0",
    # Database
    "sqlmodel>=0.0.21",
    "aiosqlite>=0.20.0",
    # Document parsing
    "PyMuPDF>=1.24.0",
    "python-docx>=1.1.0",
    # Apple integrations (macOS only)
    "pyobjc-framework-EventKit>=10.0;sys_platform=='darwin'",
    # Audio
    "sounddevice>=0.5.0",
    "soundfile>=0.12.0",
    # File utilities
    "Pillow>=10.0.0",
    "img2pdf>=0.5.0",
    "watchdog>=4.0.0",
    # General
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[project.scripts]
tp = "cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "cli*"]

[tool.ruff]
line-length = 100
target-version = "py310"
